requirements.txt
asgiref==3.6.0
cffi==1.15.1
Django==4.1.5
django-environ==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
pycparser==2.21
PyNaCl==1.5.0
pytz==2022.7.1
sqlparse==0.4.3
